This PR extends the already-existing modifications we make to Landscape
configs in `wsl-pro-service`. It appends `tags = wsl` to the Landscape
config if a `tags` key doesn't already exist. Any `tags` already present
in the config (including empty tags) are *not* overridden.

*For review, the [Landscape example
configuration](https://github.com/canonical/landscape-client/blob/master/example.conf#L163-L167)
may be useful.*

---

UDENG-2372

+++ wsl-pro-service/internal/system/landscape.go
@@ -89,11 +89,15 @@ func modifyConfig(ctx context.Context, s *System, landscapeConfig string, hostag
 	if err != nil {
 		return "", err
 	}
-	if err := overrideKey(ctx, data, "client", "computer_title", distroName); err != nil {
+	if err := createKey(ctx, data, "client", "computer_title", distroName, true); err != nil {
 		return "", err
 	}
 
-	if err := overrideKey(ctx, data, "client", "hostagent_uid", hostagentUID); err != nil {
+	if err := createKey(ctx, data, "client", "hostagent_uid", hostagentUID, true); err != nil {
+		return "", err
+	}
+
+	if err := createKey(ctx, data, "client", "tags", "wsl", false); err != nil {
 		return "", err
 	}
 
@@ -109,8 +113,8 @@ func modifyConfig(ctx context.Context, s *System, landscapeConfig string, hostag
 	return w.String(), nil
 }
 
-// overrideKey sets a key to a particular value.
-func overrideKey(ctx context.Context, data *ini.File, section, key, value string) error {
+// createKey tries to create a key with a particular value, optionally overriding an existing key.
+func createKey(ctx context.Context, data *ini.File, section, key, value string, override bool) error {
 	sec, err := data.GetSection(section)
 	if err != nil {
 		if sec, err = data.NewSection(section); err != nil {
@@ -119,6 +123,11 @@ func overrideKey(ctx context.Context, data *ini.File, section, key, value string
 	}
 
 	if sec.HasKey(key) {
+		if !override {
+			log.Infof(ctx, "Landscape config contains key %q. Its value will not be overridden.", key)
+			return nil
+		}
+
 		log.Infof(ctx, "Landscape config contains key %q. Its value will be overridden.", key)
 		sec.DeleteKey(key)
 	}
